Added exceptions, noexcept, nodiscard, more API, better logic, changed int to size_t (+tests)

+++ tests/test_sequence/main_sequence_test.cpp
@@ -1,8 +1,8 @@
-int testBasics();
-int testOperators();
+size_t testBasics();
+size_t testOperators();
 
-int testSequence() {
-	int passedTests = 0;
+size_t testSequence() {
+	size_t passedTests = 0;
 	passedTests += testBasics();
 	passedTests += testOperators();
 	return passedTests;
